Added settings
