+++ src/app/components/settings/danger-zone/danger-zone.component.html
@@ -1,0 +1,10 @@
+<div fxLayoutAlign="center center" fxFlexFill class="content">
+  <mat-card fxFlex="25">
+    <mat-toolbar color="warn">Danger Zone</mat-toolbar>
+    <form fxLayoutAlign="stretch" fxLayout="column" class="login-form">
+      <button mat-raised-button class="mt-5" color="warn" (click)="openFlagDialog()">
+            Delete Account
+      </button>
+    </form>
+  </mat-card>
+</div>
