+++ src/app/components/settings/danger-zone/danger-zone.component.scss
@@ -1,0 +1,22 @@
+.content {
+    margin-top: 5%;
+  }
+  
+  .content mat-card, .notLoggedIn mat-card {
+    padding: 0;
+    max-width: 500px !important;
+    margin: 5px;
+  }
+  
+  .login-form {
+    padding: 20px  ;
+  }
+
+  button {
+    margin: 5px;
+  }
+  
+  .form-button {
+    margin: 1.5em 5px;
+  }
+
