+++ src/app/components/settings/settings.component.html
@@ -1,0 +1,6 @@
+<div class="logout-div">
+    <button mat-raised-button class="mt-5" color="accent" (click)="logOut()">Log out</button>
+</div>
+
+<app-user-settings></app-user-settings>
+<app-danger-zone></app-danger-zone>
