+++ src/app/components/settings/settings.component.scss
@@ -1,0 +1,3 @@
+.logout-div {
+    margin: 1%;
+}
