+++ src/app/components/user-settings/user-settings.component.html
@@ -20,11 +20,6 @@
       </label>
 
         <button mat-raised-button type="submit" class="form-button mt-5" color="primary">Save</button>
-
-        <button mat-raised-button class="mt-5" color="accent" (click)="logOut()">Log out</button>
-
-      <button mat-raised-button class="mt-5" color="warn" (click)="openFlagDialog()">Delete User</button>
-      <a id="request-data" (click)="openSnackBar()">Request data</a>
     </form>
   </mat-card>
 </div>
