+++ src/app/routing/app-routing.module.ts
@@ -1,11 +1,10 @@
+import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
 // Components
 import { HomeComponent } from '../components/home/home.component';
 import { LoginComponent } from '../components/login/login.component';
 import { NewPostComponent } from '../components/newpost/newpost.component';
-import { UserSettingsComponent } from '../components/user-settings/user-settings.component';
-import { NgModule } from '@angular/core';
 import { EditpostComponent } from '../components/editpost/editpost.component';
 import { RegisterComponent } from '../components/register/register.component';
 import { NotfoundComponent } from "../components/notfound/notfound.component";
@@ -15,11 +14,12 @@ import { TermsComponent } from '../components/terms/terms.component';
 import { VerifyComponent } from "../components/verify/verify.component";
 import { ResetComponent} from "../components/reset/reset.component";
 import { ForgotComponent} from "../components/forgot/forgot.component";
+import { SettingsComponent } from '../components/settings/settings.component';
 
 export const routes: Routes = [
   { path: '', component: HomeComponent, data: { animation: 'HomePage' } },
   { path: 'login', component: LoginComponent, data: { animation: 'LoginPage' } },
-  { path: 'user', component: UserSettingsComponent, data: { animation: 'UserPage' } },
+  { path: 'user', component: SettingsComponent, data: { animation: 'UserPage' } },
   { path: 'editpost/:post_id', component: EditpostComponent, data: { animation: 'EditPostPage' } },
   { path: 'newpost', component: NewPostComponent, data: { animation: 'NewPostPage' } },
   { path: 'register', component: RegisterComponent, data: { animation: 'RegisterPage' } },
